Assert that a data version in a data package matches an expectation.

Assert that a data version in a data package matches an expectation.

assert_data_version(
  data_package_name = NULL,
  version_string = NULL,
  acceptable = "equal",
  ...
)

Arguments

data_package_name
character Name of the package.

version_string
character Version string in "x.y.z" format.

acceptable
character one of "equal", "equal_or_greater", describing what version match is acceptable.

...
additional arguments passed to data_version (such as lib.loc)

Value

invisible logical TRUE if success, otherwise stop on mismatch.

Details

Tests the DataVersion string in data_package_name against version_string testing the major, minor and revision portion.

Tests "data_package_name version equal version_string" or "data_package_name version equal_or_greater version_string".

Examples

if(rmarkdown::pandoc_available()){
f <- tempdir()
f <- file.path(f, "foo.Rmd")
con <- file(f)
writeLines("```{r}\n tbl = table(sample(1:10,1000,replace=TRUE)) \n```\n",con = con)
close(con)
pname <- basename(tempfile())
datapackage_skeleton(name = pname,
   path=tempdir(),
   force = TRUE,
   r_object_names = "tbl",
   code_files = f)
package_build(file.path(tempdir(),pname), install = FALSE)

devtools::load_all(file.path(tempdir(),pname))

assert_data_version(data_package_name = pname,version_string = "0.1.0",acceptable = "equal")
}

#> ✔ Creating '/tmp/Rtmpvex5ls/file4404455c9fc/'

#> ✔ Setting active project to '/tmp/Rtmpvex5ls/file4404455c9fc'

#> ✔ Creating 'R/'

#> ✔ Writing 'DESCRIPTION'

#> Package: file4404455c9fc
#> Title: What the Package Does (One Line, Title Case)
#> Version: 0.0.0.9000
#> Authors@R (parsed):
#>     * First Last <first.last@example.com> [aut, cre] (YOUR-ORCID-ID)
#> Description: What the package does (one paragraph).
#> License: `use_mit_license()`, `use_gpl3_license()` or friends to
#>     pick a license
#> Encoding: UTF-8
#> LazyData: true
#> Roxygen: list(markdown = TRUE)
#> RoxygenNote: 7.1.1

#> ✔ Writing 'NAMESPACE'

#> ✔ Setting active project to '/tmp/Rtmpvex5ls/file44043904168'

#> ✔ Setting active project to '/tmp/Rtmpvex5ls/file4404455c9fc'

#> ✔ Added DataVersion string to 'DESCRIPTION'

#> ✔ Creating 'data-raw/'

#> ✔ Creating 'data/'

#> ✔ Creating 'inst/extdata/'

#> ✔ Copied foo.Rmd into 'data-raw'
#> ✔ configured 'datapackager.yml' file
#> 
#> ✔ 1 data set(s) created by foo.Rmd
#> • tbl
#> ☘ Built  all datasets!
#> Non-interactive NEWS.md file update.
#> 

#> ✔ Creating 'vignettes/'

#> ✔ Creating 'inst/doc/'

#> First time using roxygen2. Upgrading automatically...

#> ℹ Loading file4404455c9fc

#> Writing NAMESPACE
#> Writing file4404455c9fc.Rd
#> Writing tbl.Rd
#>      checking for file ‘/tmp/Rtmpvex5ls/file4404455c9fc/DESCRIPTION’ ...  ✔  checking for file ‘/tmp/Rtmpvex5ls/file4404455c9fc/DESCRIPTION’
#>   ─  preparing ‘file4404455c9fc’:
#>    checking DESCRIPTION meta-information ...  ✔  checking DESCRIPTION meta-information
#>   ─  checking for LF line-endings in source and make files and shell scripts
#>   ─  checking for empty or unneeded directories
#>        NB: this package now depends on R (>= 3.5.0)
#>        WARNING: Added dependency on R >= 3.5.0 because serialized objects in
#>      serialize/load version 3 cannot be read in older versions of R.
#>      File(s) containing such objects:
#>        ‘file4404455c9fc/data/tbl.rda’
#>   ─  building ‘file4404455c9fc_1.0.tar.gz’
#>      
#> Next Steps 
#> 1. Update your package documentation. 
#>    - Edit the documentation.R file in the package source data-raw subdirectory and update the roxygen markup. 
#>    - Rebuild the package documentation with  document() . 
#> 2. Add your package to source control. 
#>    - Call  git init .  in the package source root directory. 
#>    -  git add  the package files. 
#>    -  git commit  your new package. 
#>    - Set up a github repository for your pacakge. 
#>    - Add the github repository as a remote of your local package repository. 
#>    -  git push  your local repository to gitub. 

#> ℹ Loading file4404455c9fc