Get the DataVersion for a package

Retrieves the DataVersion of a package if available

data_version(pkg, lib.loc = NULL)

dataVersion(pkg, lib.loc = NULL)

Arguments

pkg
character the package name

lib.loc
character path to library location.

Note

dataVersion() has been renamed to data_version()

See also

Examples

if(rmarkdown::pandoc_available()){
f <- tempdir()
f <- file.path(f,"foo.Rmd")
con <- file(f)
writeLines("```{r}\n tbl = table(sample(1:10,1000,replace=TRUE)) \n```\n",con=con)
close(con)
pname <- basename(tempfile())
datapackage_skeleton(name = pname,
   path=tempdir(),
   force = TRUE,
   r_object_names = "tbl",
   code_files = f)

   package_build(file.path(tempdir(),pname), install = FALSE)

   devtools::load_all(file.path(tempdir(),pname))
   data_version(pname)
}

#> ✔ Setting active project to '/private/var/folders/jh/x0h3v3pd4dd497g3gtzsm8500000gn/T/RtmpXhvbVH/file1182814adb0e3'
#> ✔ Creating 'R/'
#> ✔ Creating 'man/'
#> ✔ Writing 'DESCRIPTION'
#> ✔ Writing 'NAMESPACE'
#> ✔ Added DataVersion string to 'DESCRIPTION'
#> ✔ Creating 'data-raw/'
#> ✔ Creating 'data/'
#> ✔ Creating 'inst/extdata/'
#> ✔ Copied foo.Rmd into 'data-raw'
#> ✔ configured 'datapackager.yml' file
#> 
#> ✔ Setting active project to '/private/var/folders/jh/x0h3v3pd4dd497g3gtzsm8500000gn/T/RtmpXhvbVH/file1182814adb0e3'
#> ✔ 1 data set(s) created by foo.Rmd
#> • tbl
#> ☘ Built  all datasets!
#> Non-interactive NEWS.md file update.
#> 
#> ✔ Creating 'vignettes/'
#> ✔ Creating 'inst/doc/'

#> First time using roxygen2. Upgrading automatically...

#> Updating roxygen version in /private/var/folders/jh/x0h3v3pd4dd497g3gtzsm8500000gn/T/RtmpXhvbVH/file1182814adb0e3/DESCRIPTION

#> Writing NAMESPACE

#> Loading file1182814adb0e3

#> Writing file1182814adb0e3.Rd
#> Writing tbl.Rd
#>      checking for file ‘/private/var/folders/jh/x0h3v3pd4dd497g3gtzsm8500000gn/T/RtmpXhvbVH/file1182814adb0e3/DESCRIPTION’ ...  ✔  checking for file ‘/private/var/folders/jh/x0h3v3pd4dd497g3gtzsm8500000gn/T/RtmpXhvbVH/file1182814adb0e3/DESCRIPTION’
#>   ─  preparing ‘file1182814adb0e3’:
#>   ✔  checking DESCRIPTION meta-information
#>   ─  checking for LF line-endings in source and make files and shell scripts
#>   ─  checking for empty or unneeded directories
#> ─  looking to see if a ‘data/datalist’ file should be added
#>        NB: this package now depends on R (>= 3.5.0)
#>        WARNING: Added dependency on R >= 3.5.0 because serialized objects in  serialize/load version 3 cannot be read in older versions of R.  File(s) containing such objects: 'file1182814adb0e3/data/tbl.rda'
#>   ─  building 'file1182814adb0e3_1.0.tar.gz'
#>      
#> Next Steps 
#> 1. Update your package documentation. 
#>    - Edit the documentation.R file in the package source data-raw subdirectory and update the roxygen markup. 
#>    - Rebuild the package documentation with  document() . 
#> 2. Add your package to source control. 
#>    - Call  git init .  in the package source root directory. 
#>    -  git add  the package files. 
#>    -  git commit  your new package. 
#>    - Set up a github repository for your pacakge. 
#>    - Add the github repository as a remote of your local package repository. 
#>    -  git push  your local repository to gitub. 

#> Loading file1182814adb0e3

#> [1] ‘0.1.0’